Merged PR 1: Added CAN functionality to the devkit and enabled the J1939 kernel module
Added what is called a kernel module fragment to the repository that defines any changes that need to be made to the linux kernel before the image is built and added an example for how that fragment is going to be added to the project through the .bbappend file in the recipe-kernel directory.

I've also added documentation to describe how the fragments are generated for the project using the menuconfig option for the kernel and how those files are supposed to be added to this repository. Long term I think this is something that we can script so that there is no possibility for messing this up.

Updated the device tree for the devkit to configure the CAN interface on the processor.