Merged PR 17: Added Touchscreen driver
Added the touchscreen driver to the distribution and included the corresponding devicetree bindings to make sure that everything is working correctly. Additionally there was a problem with touch on the 5 inch but that ended up being caused by too much logging in the kernel level.

All Touchscreens is now working perfectly on the 5inch and 7inch.

There was a small delay introduced in boot times with the driver that was added, but that has been mitigated by Jonathan adding a deferred start to the init.