Merged PR 44: Updated the base-files recipe
Updated the base-file recipe. Previously the recipe was adding a new file location to the original recipe then adding new steps to the install step that would undo all the processing that was being done on the profile to have it working properly. This has been reverted as it wasn't necessary and the recipe has been simplified to only what it needs to have.