Merged PR 153: Added network configuration files to the distribution that will configure CAN...
Added network configuration files to the distribution that will configure CAN on the board automatically to 250K baud. Additionally added a new service whose sole purpose is to help in the configuration of the system at startup. So any commands that should be run on startup to make sure that the system is properly configured for the customer can be placed there.