Merged PR 160: BSP Versioning architecture
Simple updates to allow for long term versioning of the image. Contained with in this pull request are 2 text files for storing version strings and descriptions as well as a shell script that will provide quick access to the version details. Mostly just architecture setup until a long term versioning strategy is implemented