Merged PR 162: Added device tree bindings to the image to ensure that rs485 doesn't need to...
Added device tree bindings to the image to ensure that rs485 doesn't need to be set up before it is used as a serial port. This will allow for all communication through the rs485 port to occur as if it was a regular serial device.

You can easily test this functionality out to make sure it's working by building the spa touch application without the call to `ioctl(fd, TIOCSRS485, &rs485conf)` found in the `rs485device.c` file.