Merged PR 164: Updating the Kernel Interrupt Frequency from 100HZ to 1KHZ
Updated the Kernel Interrupt Frequency from 100HZ to 1KHZ by adding a fragment to set the CONFIG_HZ variable to 1000. This was done in order to achieve better throughput on the rcd platform. The updated configuration has been tested and no side effects to user space applications have been reported.