Merged PR 194: Updated systemd timesyncd.conf with real timeservers
Updated the systemd definition for the timesyncd.conf to make sure that it is pointing to an actual time server so that the time on the device can be updated whenever it has connection to the internet.

There is also a command that you can run manually to set the time on the device in question from a unit system. That command looks something like this: `ssh root@192.168.7.1 date -s @<epoch time in seconds>`

This has been tested and confirmed to be working.