Merged PR 195: Device Tree Overlays + patches for new 7inch high res screens
This PR contains the first iteration for working with linux kernel device tree overlays.
I have started a wiki entry with info on overlays and how the current one was written / compiled / implemented. The wiki entry is still under development but can be found here: https://dev.azure.com/HeliosEngineering/Right%20Cost%20Display/_wiki/wikis/Development%20Documentation/71/Device-Tree-Overlays-for-the-RCD-BSP#
Beyond just the overlays (which were necessary with the introduction of the new 7inch high resolution new vision screen ), this pr also consolidated the weston background images and uses new higher resolution images given to me by allison. Furthermore, on the new rev 7 boards, 2 of the LCD pins are now controlled by GPIO and their default state is not the desired state. I introduce a script called lcd-config which will set the pins to the correct state and adjust the splash screen as well as the weston background. It also applies the device tree overlay if the screen is a 7 inch. These 2 things are temporary in that detection of the screen should be done in u-boot and then logic to load the right overlay should be handled in u-boot , not user space. The other temporary fix is to add the lcd-pin-config line to the init script so that the splash screen comes up in the proper orientation. We should fix this later by configuring the device tree to set the gpio pins which control the lcd U/D and L/R to their proper default state. It should be noted that older rev boards should be uneffected by these changes and I tested on several to make sure.
Other than that, I have a few TODOs which involve adding a recipe to compile the overlay and a recipe to compile the lcd-pin-config binary automatically but ran into issues with setting that up so I'll tackle it later.