Merged PR 211: U-Boot Optimizations
Updates to U-boot Boot command and defconfig to speed up u-boot. Main change is to the actual distro boot command provided by ST. The original command tries to accommodate all supported boot methods as well as all supported hardware but we don't need our boot command to be so robust since these things will likely never change for our board. Other than that, there are a few configs that are set to disabled.