Merged PR 221: Scripts to publish webkit via debian packages
Added scripts to tar the necessary Webkit Debian packages and deploy to an RCD with the base image.
This is not fool proof:
* There is no way to verify the packages are compatible with the base image to which they are applied.
* The developer has to manually build the webkit image prior to running the script.

This is intended for developers, for example for the certification test app, to be able to take a Spark released base image and update it to support the Webkit browser.